rename  RewardGraph3 to RewardGraph and delete test.tsx and test2.tsx
